Split workspace-card helpers into a new module to fix lint size limits
Extracted the workspace-card list/detail helpers from workspaces.js into
workspace-cards.js, and pulled the card-page SQL builder out of
getWorkspaceCardPage in workspace-queries.js, bringing both back under the
max-lines / max-lines-per-function lint limits.
